--- src/main/java/com/github/dakusui/floorplan/tdesc/FloorPlanRunnerParameters.java
@@ -1,8 +1,0 @@
-package com.github.dakusui.floorplan.tdesc;
-
-import com.github.dakusui.floorplan.policy.Profile;
-
-public @interface FloorPlanRunnerParameters {
-  Class<? extends TestSuiteDescriptor.Factory> descriptorFactory();
-  Class<? extends Profile.Factory> profileFactory();
-}
