--- src/main/java/com/github/dakusui/floorplan/tdesc/TestBase.java
@@ -1,4 +1,0 @@
-package com.github.dakusui.floorplan.tdesc;
-
-public class TestBase {
-}
