+++ src/main/java/com/github/dakusui/floorplan/tdesc/junit4/runner/FloorPlanRunner.java
@@ -1,7 +1,7 @@
-package com.github.dakusui.floorplan.examples.bookstore;
+package com.github.dakusui.floorplan.tdesc.junit4.runner;
 
-import com.github.dakusui.floorplan.examples.bookstore.floorplan.BookstoreProfile;
-import com.github.dakusui.floorplan.examples.bookstore.tdescs.SmokeTestDesc;
+import com.github.dakusui.floorplan.exception.Exceptions;
+import com.github.dakusui.floorplan.policy.Profile;
 import com.github.dakusui.floorplan.tdesc.TestSuiteDescriptor;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -10,14 +10,29 @@ import org.junit.runners.model.FrameworkMethod;
 import org.junit.runners.model.TestClass;
 
 import java.lang.annotation.Annotation;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
 import java.util.Collection;
 import java.util.List;
 import java.util.function.Function;
 
+import static com.github.dakusui.floorplan.utils.Checks.requireNonNull;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
 import static java.util.Arrays.asList;
 import static java.util.stream.Collectors.toList;
 
 public class FloorPlanRunner extends Parameterized {
+  @Retention(RUNTIME)
+  @Inherited
+  public @interface UseProfileFactory {
+    Class<? extends Profile.Factory> value();
+  }
+
+  @Retention(RUNTIME)
+  @Inherited
+  public @interface UseTestSuiteDescriptorFactor {
+    Class<? extends TestSuiteDescriptor.Factory> value();
+  }
 
   /**
    * Only called reflectively. Do not use programmatically.
@@ -31,7 +46,7 @@ public class FloorPlanRunner extends Parameterized {
   @Override
   protected TestClass createTestClass(Class<?> testClass) {
     return new TestClass(testClass) {
-      private final TestSuiteDescriptor descriptor = createTestSuiteDescriptor();
+      private final TestSuiteDescriptor descriptor = createTestSuiteDescriptor(testClass);
 
       @Override
       public List<FrameworkMethod> getAnnotatedMethods(Class<? extends Annotation> annotationClass) {
@@ -52,8 +67,24 @@ public class FloorPlanRunner extends Parameterized {
     };
   }
 
-  private TestSuiteDescriptor createTestSuiteDescriptor() {
-    return new SmokeTestDesc().create(new BookstoreProfile());
+  private static TestSuiteDescriptor createTestSuiteDescriptor(Class<?> testClass) {
+    return createTestSuiteDescriptorFactory(testClass).create(createProfileFactory(testClass).create());
+  }
+
+  private static TestSuiteDescriptor.Factory<?> createTestSuiteDescriptorFactory(Class<?> testClass) {
+    try {
+      return requireNonNull(testClass.getAnnotation(UseTestSuiteDescriptorFactor.class)).value().newInstance();
+    } catch (InstantiationException | IllegalAccessException e) {
+      throw Exceptions.rethrow(e);
+    }
+  }
+
+  private static Profile.Factory<?> createProfileFactory(Class<?> testClass) {
+    try {
+      return requireNonNull(testClass.getAnnotation(UseProfileFactory.class).value()).newInstance();
+    } catch (InstantiationException | IllegalAccessException e) {
+      throw Exceptions.rethrow(e);
+    }
   }
 
   @Override
